sampleproject/app/tests/factories.py
import factory
from django.contrib.auth.models import User


class UserFactory(factory.DjangoModelFactory):
    class Meta:
        model = User
    username = factory.Sequence(lambda n: 'user%d' % n)
    first_name = 'John'
    last_name = 'Doe'
