sampleproject/requirements.txt
factory_boy
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.6.4
graphene==2.0
graphene-django==2.0
-e ..